# docs/03-api-openapi.json
{
  "components": {
    "schemas": {
...
      "CategoryValueOut": {
        "properties": {
          "category": {
            "title": "Category",
            "type": "string"
          },
          "expired_value": {
            "title": "Expired Value",
            "type": "string"
          },
          "sellable_value": {
            "title": "Sellable Value",
            "type": "string"
          },
          "total_value": {
            "title": "Total Value",
            "type": "string"
          }
        },
        "required": [
          "total_value",
          "sellable_value",
          "expired_value",
          "category"
        ],
        "title": "CategoryValueOut",
        "type": "object"
      },
...
      "ExpiredLotWithValueOut": {
        "properties": {
          "days_expired": {
            "title": "Days Expired",
            "type": "integer"
          },
          "expiry_date": {
            "format": "date",
            "title": "Expiry Date",
            "type": "string"
          },
          "lot_id": {
            "format": "uuid",
            "title": "Lot Id",
            "type": "string"
          },
          "lot_number": {
            "title": "Lot Number",
            "type": "string"
          },
          "medicine_id": {
            "format": "uuid",
            "title": "Medicine Id",
            "type": "string"
          },
          "medicine_name": {
            "title": "Medicine Name",
            "type": "string"
          },
          "quantity_remaining": {
            "title": "Quantity Remaining",
            "type": "integer"
          },
          "stock_value": {
            "title": "Stock Value",
            "type": "string"
          }
        },
        "required": [
          "lot_id",
          "medicine_id",
          "medicine_name",
          "lot_number",
          "quantity_remaining",
          "expiry_date",
          "days_expired",
          "stock_value"
        ],
        "title": "ExpiredLotWithValueOut",
        "type": "object"
      },
      "ExpiringLotWithValueOut": {
        "properties": {
          "days_remaining": {
            "title": "Days Remaining",
            "type": "integer"
          },
          "expiry_date": {
            "format": "date",
            "title": "Expiry Date",
            "type": "string"
          },
          "lot_id": {
            "format": "uuid",
            "title": "Lot Id",
            "type": "string"
          },
          "lot_number": {
            "title": "Lot Number",
            "type": "string"
          },
          "medicine_id": {
            "format": "uuid",
            "title": "Medicine Id",
            "type": "string"
          },
          "medicine_name": {
            "title": "Medicine Name",
            "type": "string"
          },
          "quantity_remaining": {
            "title": "Quantity Remaining",
            "type": "integer"
          },
          "risk_level": {
            "title": "Risk Level",
            "type": "string"
          },
          "stock_value": {
            "title": "Stock Value",
            "type": "string"
          }
        },
        "required": [
          "lot_id",
          "medicine_id",
          "medicine_name",
          "lot_number",
          "quantity_remaining",
          "expiry_date",
          "days_remaining",
          "risk_level",
          "stock_value"
        ],
        "title": "ExpiringLotWithValueOut",
        "type": "object"
      },
      "ExpiringReportWithValueOut": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExpiringLotWithValueOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "summary": {
            "additionalProperties": {
              "$ref": "#/components/schemas/RiskSummaryWithValueOut"
            },
            "title": "Summary",
            "type": "object"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset",
          "summary"
        ],
        "title": "ExpiringReportWithValueOut",
        "type": "object"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "title": "Detail",
            "type": "array"
          }
        },
        "title": "HTTPValidationError",
        "type": "object"
      },
      "InventoryValueOut": {
        "properties": {
          "by_category": {
            "items": {
              "$ref": "#/components/schemas/CategoryValueOut"
            },
            "title": "By Category",
            "type": "array"
          },
          "by_medicine": {
            "$ref": "#/components/schemas/Page_MedicineValueOut_"
          },
          "expired_value": {
            "title": "Expired Value",
            "type": "string"
          },
          "sellable_value": {
            "title": "Sellable Value",
            "type": "string"
          },
          "total_value": {
            "title": "Total Value",
            "type": "string"
          }
        },
        "required": [
          "total_value",
          "sellable_value",
          "expired_value",
          "by_medicine",
          "by_category"
        ],
        "title": "InventoryValueOut",
        "type": "object"
      },
...
      "LowStockRowOut": {
        "properties": {
          "available_quantity": {
            "title": "Available Quantity",
            "type": "integer"
          },
          "medicine_id": {
            "format": "uuid",
            "title": "Medicine Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reorder_point": {
            "title": "Reorder Point",
            "type": "integer"
          },
          "shortage": {
            "title": "Shortage",
            "type": "integer"
          }
        },
        "required": [
          "medicine_id",
          "name",
          "available_quantity",
          "reorder_point",
          "shortage"
        ],
        "title": "LowStockRowOut",
        "type": "object"
      },
...
      "MedicineValueOut": {
        "properties": {
          "expired_value": {
            "title": "Expired Value",
            "type": "string"
          },
          "medicine_id": {
            "format": "uuid",
            "title": "Medicine Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "sellable_value": {
            "title": "Sellable Value",
            "type": "string"
          },
          "total_value": {
            "title": "Total Value",
            "type": "string"
          }
        },
        "required": [
          "total_value",
          "sellable_value",
          "expired_value",
          "medicine_id",
          "name"
        ],
        "title": "MedicineValueOut",
        "type": "object"
      },
...
      "Page_ExpiredLotWithValueOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/ExpiredLotWithValueOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[ExpiredLotWithValueOut]",
        "type": "object"
      },
...
      "Page_LowStockRowOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/LowStockRowOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[LowStockRowOut]",
        "type": "object"
      },
...
      "Page_MedicineValueOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/MedicineValueOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[MedicineValueOut]",
        "type": "object"
      },
...
      "Page_StockRowWithValueOut_": {
        "properties": {
          "items": {
            "items": {
              "$ref": "#/components/schemas/StockRowWithValueOut"
            },
            "title": "Items",
            "type": "array"
          },
          "limit": {
            "title": "Limit",
            "type": "integer"
          },
          "offset": {
            "title": "Offset",
            "type": "integer"
          },
          "total": {
            "title": "Total",
            "type": "integer"
          }
        },
        "required": [
          "items",
          "total",
          "limit",
          "offset"
        ],
        "title": "Page[StockRowWithValueOut]",
        "type": "object"
      },
...
      "RiskSummaryWithValueOut": {
        "properties": {
          "lot_count": {
            "title": "Lot Count",
            "type": "integer"
          },
          "stock_value": {
            "title": "Stock Value",
            "type": "string"
          }
        },
        "required": [
          "lot_count",
          "stock_value"
        ],
        "title": "RiskSummaryWithValueOut",
        "type": "object"
      },
...
      "StockRowWithValueOut": {
        "properties": {
          "available_quantity": {
            "title": "Available Quantity",
            "type": "integer"
          },
          "available_value": {
            "title": "Available Value",
            "type": "string"
          },
          "category": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Category"
          },
          "expired_quantity": {
            "title": "Expired Quantity",
            "type": "integer"
          },
          "expired_value": {
            "title": "Expired Value",
            "type": "string"
          },
          "medicine_id": {
            "format": "uuid",
            "title": "Medicine Id",
            "type": "string"
          },
          "name": {
            "title": "Name",
            "type": "string"
          },
          "reorder_point": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reorder Point"
          },
          "strength": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Strength"
          }
        },
        "required": [
          "medicine_id",
          "name",
          "strength",
          "category",
          "available_quantity",
          "expired_quantity",
          "reorder_point",
          "available_value",
          "expired_value"
        ],
        "title": "StockRowWithValueOut",
        "type": "object"
      },
...
      "ValidationError": {
        "properties": {
          "ctx": {
            "title": "Context",
            "type": "object"
          },
          "input": {
            "title": "Input"
          },
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "title": "Location",
            "type": "array"
          },
          "msg": {
            "title": "Message",
            "type": "string"
          },
          "type": {
            "title": "Error Type",
            "type": "string"
          }
        },
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError",
        "type": "object"
      }
    },
    "securitySchemes": {
      "HTTPBearer": {
        "bearerFormat": "JWT",
        "scheme": "bearer",
        "type": "http"
      }
    }
  },
  "info": {
    "title": "AI Pharmacy Management API",
    "version": "0.1.0"
  },
  "openapi": "3.1.0",
  "paths": {
...
    "/api/v1/reports/expired": {
      "get": {
        "description": "staff: stock_value omitted.",
        "operationId": "expired_api_v1_reports_expired_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_ExpiredLotWithValueOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Expired",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/expiring": {
      "get": {
        "description": "staff: stock_value omitted.",
        "operationId": "expiring_api_v1_reports_expiring_get",
        "parameters": [
          {
            "in": "query",
            "name": "days",
            "required": false,
            "schema": {
              "default": 180,
              "maximum": 3650,
              "minimum": 1,
              "title": "Days",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/ExpiringReportWithValueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Expiring",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/inventory-value": {
      "get": {
        "operationId": "inventory_value_api_v1_reports_inventory_value_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/InventoryValueOut"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Inventory Value",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/low-stock": {
      "get": {
        "operationId": "low_stock_api_v1_reports_low_stock_get",
        "parameters": [
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_LowStockRowOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Low Stock",
        "tags": [
          "reports"
        ]
      }
    },
    "/api/v1/reports/stock": {
      "get": {
        "description": "staff: *_value omitted.",
        "operationId": "stock_api_v1_reports_stock_get",
        "parameters": [
          {
            "in": "query",
            "name": "q",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Q"
            }
          },
          {
            "in": "query",
            "name": "category",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "maxLength": 200,
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "title": "Category"
            }
          },
          {
            "in": "query",
            "name": "limit",
            "required": false,
            "schema": {
              "default": 50,
              "maximum": 200,
              "minimum": 1,
              "title": "Limit",
              "type": "integer"
            }
          },
          {
            "in": "query",
            "name": "offset",
            "required": false,
            "schema": {
              "default": 0,
              "minimum": 0,
              "title": "Offset",
              "type": "integer"
            }
          }
        ],
        "responses": {
          "200": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/Page_StockRowWithValueOut_"
                }
              }
            },
            "description": "Successful Response"
          },
          "422": {
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            },
            "description": "Validation Error"
          }
        },
        "security": [
          {
            "HTTPBearer": []
          }
        ],
        "summary": "Stock",
        "tags": [
          "reports"
        ]
      }
    },
...
  }
}